move the customise work out into a helper function
